requirements.txt
fastapi
uvicorn[standard]
python-multipart
faster-whisper
python-dotenv